dashboard/dashboard_requirements.txt
"""
requirements.txt 
    Required dependencies for the Dash application
    This file lists the necessary Python libraries and their versions needed to run the dashboard.

Author:   Eyerusalem Mamuye Birhan
Date:     2025-03-13
Version:  1.0

Copyright (C) 2023 CSSR4Africa Consortium
This project is funded by the African Engineering and Technology Network (Afretec)
Inclusive Digital Transformation Research Grant Programme.
Website: www.cssr4africa.org

This program comes with ABSOLUTELY NO WARRANTY.
"""
dash>=2.15.0
plotly>=5.19.0
pandas
gspread>=6.0.0
google-auth>=2.25.0
openpyxl>=2.1.3
